requirements.txt
flask==3.0.2
flask-wtf==1.2.1
instagrapi==2.0.0
python-dotenv==1.0.0
rich==13.7.0
werkzeug==3.0.1
wtforms==3.1.2 